PR-FIX: only create topic subscription on message pump on Service Bus topic

+++ src/Arcus.Messaging.Pumps.ServiceBus/AzureServiceBusMessagePump.cs
@@ -68,7 +68,8 @@ namespace Arcus.Messaging.Pumps.ServiceBus
         /// <param name="cancellationToken">Indicates that the start process has been aborted.</param>
         public override async Task StartAsync(CancellationToken cancellationToken)
         {
-            if (Settings.Options.TopicSubscription.HasFlag(TopicSubscription.CreateOnStart))
+            if (Settings.ServiceBusEntity == ServiceBusEntity.Topic
+                && Settings.Options.TopicSubscription.HasFlag(TopicSubscription.CreateOnStart))
             {
                 await CreateTopicSubscriptionAsync(cancellationToken);
             }
@@ -312,7 +313,8 @@ namespace Arcus.Messaging.Pumps.ServiceBus
         /// <param name="cancellationToken">Indicates that the shutdown process should no longer be graceful.</param>
         public override async Task StopAsync(CancellationToken cancellationToken)
         {
-            if (Settings.Options.TopicSubscription.HasFlag(TopicSubscription.DeleteOnStop))
+            if (Settings.ServiceBusEntity == ServiceBusEntity.Topic
+                && Settings.Options.TopicSubscription.HasFlag(TopicSubscription.DeleteOnStop))
             {
                 await DeleteTopicSubscriptionAsync(cancellationToken);
             }
